rename: 구독버튼 index.js로 변경
