+++ src/components/main/list/List.js
@@ -9,7 +9,7 @@ import { useGetAtom } from "../../../store/atom.js";
 import { PRESS_ICON } from "../../../constants/constants.js";
 import { _querySelector } from "../../../utils/my-query-selector.js";
 import { checkIsAllType, checkIsGridView } from "../../../utils/utils.js";
-import { getSubscribeButton } from "../../common/subscribe-button/SubscribeButton.js";
+import { getSubscribeButton } from "../../common/subscribe-button/index.js";
 
 const $listView = _querySelector(".list-view_main");
